+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Update attributes endpoint to V3
 - Add changelog
+- Leverage payload attributes post-verification